+++ Sources/VComponents/Components/Containers/Disclosure Group/VDisclosureGroupUIModel.swift
@@ -29,8 +29,8 @@ public struct VDisclosureGroupUIModel {
     /// Corner radii . Set to to `15`s.
     public var cornerRadii: RectangleCornerRadii = .init(15)
 
-    /// Indicates if left and right corners should switch to support RTL languages. Set to `true`.
-    public var reversesLeftAndRightCornersForRTLLanguages: Bool = true
+    /// Indicates if horizontal corners should switch to support RTL languages. Set to `true`.
+    public var reversesHorizontalCornersForRTLLanguages: Bool = true
 
     // MARK: Properties - Background
     /// Background color.
@@ -48,7 +48,7 @@ public struct VDisclosureGroupUIModel {
         var uiModel: VGroupBoxUIModel = .init()
 
         uiModel.cornerRadii = cornerRadii
-        uiModel.reversesLeftAndRightCornersForRTLLanguages = reversesLeftAndRightCornersForRTLLanguages
+        uiModel.reversesHorizontalCornersForRTLLanguages = reversesHorizontalCornersForRTLLanguages
 
         uiModel.backgroundColor = backgroundColor