+++ Sources/VComponents/Components/Containers/Group Box/VGroupBoxUIModel.swift
@@ -15,8 +15,8 @@ public struct VGroupBoxUIModel {
     /// Corner radii . Set to to `15`s.
     public var cornerRadii: RectangleCornerRadii = .init(15)
 
-    /// Indicates if left and right corners should switch to support RTL languages. Set to `true`.
-    public var reversesLeftAndRightCornersForRTLLanguages: Bool = true
+    /// Indicates if horizontal corners should switch to support RTL languages. Set to `true`.
+    public var reversesHorizontalCornersForRTLLanguages: Bool = true
 
     // MARK: Properties - Background
     /// Background color.