+++ Sources/VComponents/Components/Indicators (Definite)/Progress Bar/VProgressBarUIModel.swift
@@ -56,10 +56,8 @@ public struct VProgressBarUIModel {
 #endif
     }()
 
-    /// Indicates if progress bar rounds progress view right-edge. Set to `true`.
-    ///
-    /// For RTL languages, this refers to left-edge.
-    public var roundsProgressViewRightEdge: Bool = true
+    /// Indicates if progress bar rounds progress view trailing corners. Set to `true`.
+    public var roundsProgressViewTrailingCorners: Bool = true
 
     // MARK: Properties - Track
     /// Track color.