+++ Sources/VComponents/Components/Modals/Alert/VAlertUIModel.swift
@@ -38,8 +38,8 @@ public struct VAlertUIModel {
     /// Corner radii . Set to to `20`s.
     public var cornerRadii: RectangleCornerRadii = .init(20)
 
-    /// Indicates if left and right corners should switch to support RTL languages. Set to `true`.
-    public var reversesLeftAndRightCornersForRTLLanguages: Bool = true
+    /// Indicates if horizontal corners should switch to support RTL languages. Set to `true`.
+    public var reversesHorizontalCornersForRTLLanguages: Bool = true
 
     // MARK: Properties - Background
     /// Background color.
@@ -55,7 +55,7 @@ public struct VAlertUIModel {
         var uiModel: VGroupBoxUIModel = .init()
 
         uiModel.cornerRadii = cornerRadii
-        uiModel.reversesLeftAndRightCornersForRTLLanguages = reversesLeftAndRightCornersForRTLLanguages
+        uiModel.reversesHorizontalCornersForRTLLanguages = reversesHorizontalCornersForRTLLanguages
 
         uiModel.backgroundColor = backgroundColor