+++ Sources/VComponents/Extensions/RectangleCornerRadii.HorizontalCornersReversed.swift
@@ -1,20 +1,20 @@
 //
-//  RectangleCornerRadii.WithReversedLeftAndRightCorners.swift
-//
+//  RectangleCornerRadii.HorizontalCornersReversed.swift
+//  VComponents
 //
 //  Created by Vakhtang Kontridze on 09.07.24.
 //
 
 import SwiftUI
 
-// MARK: - Rectangle Corner Radii with Reversed Left and Right Corners
+// MARK: - Rectangle Corner Radii with Horizontal Corners Reversed
 extension RectangleCornerRadii {
-    func withReversedLeftAndRightCorners(
-        _ condition: Bool = true
+    func horizontalCornersReversed(
+        if condition: Bool = true
     ) -> Self {
         guard condition else { return self }
 
-        return Self(
+        return RectangleCornerRadii(
             topLeading: topTrailing,
             bottomLeading: bottomTrailing,
             bottomTrailing: bottomLeading,